Add embedded Python interpreter (unsuccessfully)
While we can build alright, `libpython2.7.so` isn't linked to our
lib correctly, so when calling the embedded Python interpreter from
Java, a linkage error is thrown. (This is true for both linking in
`ibpython` as STATIC and SHARED lib).

The error message seen via `logcat` indicates that `libpython` is
expected in a different directory (`./obj/local/armeabi/libpython2.7.so`)
than where the app has it (`/data/data/com.sensibility_testbed/lib`).
The same wrong path is found in the ELF object of our library.
I'm researching where else to put `libpython`, or alternatively how
link it so that it will be found where it is. (From what I've read,
just pointing `LD_LIBRARY_PATH` to it will work, but this workaround
is not recommended for dealing with library paths from within an app....)